requirements.txt
asgiref              == 3.7.2
certifi              == 2024.2.2
charset-normalizer   == 3.3.2
Django               == 5.0.1
django-froala-editor == 4.1.4
idna                 == 3.6
packaging            == 23.2
pip                   ==23.2.1
platformdirs          ==4.1.0
psycopg              == 3.1.17
psycopg-binary       == 3.1.17
psycopg2             == 2.9.9
psycopg2-binary      == 2.9.9
requests             == 2.31.0
sqlparse             == 0.4.4
typing_extensions    == 4.9.0
tzdata               == 2023.4
urllib3              == 2.2.0
